requirements.txt
appnope==0.1.0
argon2-cffi==20.1.0
attrs==20.1.0
backcall==0.2.0
beautifulsoup4==4.9.1
bleach==3.1.5
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
diffimg==0.2.3
entrypoints==0.3
filelock==3.0.12
future==0.18.2
idna==2.10
ipykernel==5.3.4
ipython==7.17.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
joblib==0.16.0
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
jupyterlab==2.2.6
jupyterlab-server==1.2.0
kiwisolver==1.2.0
lmdb==0.99
MarkupSafe==1.1.1
matplotlib==3.3.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.1.3
numpy==1.19.1
opencv-python==4.4.0.42
packaging==20.4
pandocfilters==1.4.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
prometheus-client==0.8.0
prompt-toolkit==3.0.6
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
python-Levenshtein==0.12.0
pyzmq==19.0.2
regex==2020.7.14
requests==2.24.0
sacremoses==0.0.43
Send2Trash==1.5.0
sentencepiece==0.1.91
six==1.15.0
soupsieve==2.0.1
terminado==0.8.3
testpath==0.4.4
textdistance==4.2.0
tokenizers==0.8.1rc1
torch==1.6.0
torchvision==0.7.0
tornado==6.0.4
tqdm==4.48.2
traitlets==4.3.3
transformers==3.0.2
trdg==1.6.0
urllib3==1.25.10
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
